[Graphs] Updated code per code review comments. Branch is still WIP.
