+++ src/data-browser/graphs/quarterly/SectionFilerInfo.jsx
@@ -1,7 +1,6 @@
-import React from 'react'
-import QuarterlyFilersTable from './QuarterlyFilersTable'
+import React from "react"
+import QuarterlyFilersTable from "./QuarterlyFilersTable"
 
-export const SectionFilerInfo = ({ show }) => {
-  if (!show) return null
+export const SectionFilerInfo = () => {
   return <QuarterlyFilersTable />
 }
