+++ src/data-browser/graphs/quarterly/SectionGraphs.jsx
@@ -32,7 +32,7 @@ export const SectionGraphs = ({
   error,
   setError,
   setGraphHeaderOverview,
-  show,
+
   ...props
 }) => {
   const dispatch = useDispatch()
@@ -162,7 +162,6 @@ export const SectionGraphs = ({
     !seriesForURL ||
     !graphMenuOptions?.length
 
-  if (!show) return null
   if (loadingGraphDetails) return <LoadingIcon />
 
   return (
